WIP haskell: project euler problem 3
All stolen code really... besides `primeFactor`

I'm not satified with this solution

I believe there's a better way of doing it.

What if the largest prime of this number
is not in the first 1000 primes?

Well, then I'm wrong.

Perhaps generating the primes is not a good way of going about it.